+++ app/index.tsx
@@ -13,8 +13,8 @@ import OB1SignUp from "./(onboarding)/OB1SignUp";
 import TaskForm from "./TaskForm";
 import OB2Birthday from "./(onboarding)/OB2Birthday";
 import OB3SleepDurationGoal from "./(onboarding)/OB3SleepDurationGoal";
-import OB4SleepTime from "./(onboarding)/OB4SleepTime";
-import OB5Alarm from "./(onboarding)/OB5Alarm";
+import OB4SleepTime from "./(onboarding)/SleepTime";
+import OB5Alarm from "./(onboarding)/OB4Alarm";
 
 const Stack = createStackNavigator();
 
@@ -63,63 +63,17 @@ const Index: React.FC = () => {
   }
 
   return (
-    // <Stack.Navigator
-    //   initialRouteName="Loading"
-    //   screenOptions={{ headerShown: false, header: () => null }}
-    // >
     <View style={{ flex: 1 }}>
       {currentUser === null ? (
-        // <Stack.Screen
-        //   name="Loading"
-        //   component={LoadingScreen}
-        //   options={{ headerShown: false }}
-        // /> // Assuming you have a Loading component
-        // <Stack.Screen name="Login" component={Login} options={{ headerShown: false }} />
+        // user does not exist. Login
         <Login />
       ) : currentUserIsNew ? (
-        // <>
-        //   {/* <Stack.Screen name="OnboardingStack" component={OBLayout} /> */}
-        //   <Stack.Screen
-        //     name="OB2Birthday"
-        //     component={OB2Birthday}
-        //     options={{ headerShown: false, header: () => null }}
-        //   />
-        //   <Stack.Screen
-        //     name="OB3SleepDurationGoal"
-        //     component={OB3SleepDurationGoal}
-        //     options={{ headerShown: false, header: () => null }}
-        //   />
-        //   <Stack.Screen
-        //     name="OB4SleepTime"
-        //     component={OB4SleepTime}
-        //     options={{ headerShown: false, header: () => null }}
-        //   />
-        //   <Stack.Screen
-        //     name="OB5Alarm"
-        //     component={OB5Alarm}
-        //     options={{ headerShown: false, header: () => null }}
-        //   />
-        // </>
+        // user exists and user is new. Onboard.
         <OB2Birthday />
       ) : (
-        // <>
-        //   <Stack.Screen name="Home" component={Home} options={{ headerShown: false }} />
-        //   {/* <TabLayout /> */}
-        //   <Stack.Screen
-        //     name="TabLayout"
-        //     component={TabLayout}
-        //     options={{ headerShown: false }}
-        //   />
-        //   <Stack.Screen
-        //     name="TaskForm"
-        //     component={TaskForm}
-        //     options={{ headerShown: false }}
-        //   />
-        //   {/* <Stack.Screen name="Login" component={Login} options={{ headerShown: false }} /> */}
-        // </>
+        // user exists and user is not new. Home Screen
         <Home />
       )}
-      {/* </Stack.Navigator> */}
     </View>
   );
 };
